templates 1.0 registration

+++ src/main/java/com/taiton/dao/RoleDao.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 /**
  * Created by Taiton on 11/23/2016.
  */
-public interface RoleDao extends JpaRepository<RoleEntity, Long> {
+public interface RoleDao extends JpaRepository<RoleEntity, Integer> {
     RoleEntity findByNameRole(String name);
 }
